requirements.txt
django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.15.0
djoser==2.3.1
faker==37.4.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
supabase==2.16.0
supafunc==0.10.1
whitenoise==6.9.0
